requirements.txt
fastapi
uvicorn
deepgram-sdk==3.4.0
yt-dlp @ git+https://github.com/yt-dlp/yt-dlp.git
pydantic
ffmpeg-python
python-multipart
pytubefix
